🚀 Refactor: Improve MultiSelectDropDown implementation
- Update HomeScreen to use MultiSelectDropDownView with controller options
- Update HomeController to handle selectedOptions and options
- Remove unused MultiSelectDropDownController and related bindings
