+++ lib/domain/bindings/home_bindings.dart
@@ -1,4 +1,3 @@
-import 'package:control_system/domain/controllers/dropdown_controller.dart';
 import 'package:get/get.dart';
 
 import '../controllers/home_controller.dart';
@@ -7,6 +6,5 @@ class HomeBindings extends Bindings {
   @override
   void dependencies() {
     Get.lazyPut<HomeController>(() => HomeController());
-    Get.lazyPut<MultiSelectDropDownController>(() => MultiSelectDropDownController());
   }
 }
